Create Vending Machine
This vending machine will work as follows. A customer will first deposit an initial amount of money and then select an item from the vending machine. The program checks that the customer makes a valid selection. If it is a valid selection, then the price of selected item will be deducted from the initial deposit and also the number of items in the vending machine will be updated accordingly